requirements.txt
diskcache==5.6.3
numpy==1.26.2
openpyxl==3.1.5
pandas==2.0.3
python-dateutil==2.8.2
pytz==2023.3.post1
sqlalchemy==2.0.36